# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CU Faculty Backend API
  version: "1.0.0"
  description: >
    REST API for the CU Faculty student hub. All endpoints under `/api` require a
    valid `bff.cookie` session cookie, which is validated against the Central
    University (CU) upstream on every request. The authenticated CU student id is
    used as the local user id. The only unauthenticated endpoint is `/health`.


    List endpoints are paginated and return a `{ data, total, limit, offset }`
    envelope. Pagination is controlled with the `limit` (default 20, max 100) and
    `offset` (default 0) query parameters.
security:
  - cookieAuth: []
...
paths:
...
  /api/me:
    get:
      tags: [Me]
      summary: Get the current user
      operationId: getMe
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [Me]
      summary: Update the current user's profile
      description: >
        Accepts multipart/form-data with an optional JSON `data` part (profile
        fields) and an optional `photo` image part. Supplying `data` updates the
        profile fields; supplying `photo` sets (or replaces) the user's photo.
        Either or both parts may be omitted.
      operationId: updateMe
      requestBody:
        required: false
        description: >
          Actually sent as `multipart/form-data`: a JSON `data` part matching the
          schema below, plus an optional `photo` image part (Content-Type must
          start with `image/`) that sets or replaces the photo. Documented as
          `application/json` so the `data` schema renders; only the `data` part is
          shown here.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileData" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/me/key-skills:
    get:
      tags: [Me]
      summary: List my key skills
      operationId: getMyKeySkills
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged skills
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/key-skills/{skillId}:
    post:
      tags: [Me]
      summary: Add a key skill to me
      operationId: addMyKeySkill
      parameters:
        - $ref: "#/components/parameters/SkillId"
      responses:
        "201":
          description: Added skill
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Me]
      summary: Remove a key skill from me
      operationId: deleteMyKeySkill
      parameters:
        - $ref: "#/components/parameters/SkillId"
      responses:
        "204": { description: Removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/soft-skills:
    get:
      tags: [Me]
      summary: List my soft skills
      operationId: getMySoftSkills
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged skills
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/soft-skills/{skillId}:
    post:
      tags: [Me]
      summary: Add a soft skill to me
      operationId: addMySoftSkill
      parameters:
        - $ref: "#/components/parameters/SkillId"
      responses:
        "201":
          description: Added skill
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Skill" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Me]
      summary: Remove a soft skill from me
      operationId: deleteMySoftSkill
      parameters:
        - $ref: "#/components/parameters/SkillId"
      responses:
        "204": { description: Removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bff.cookie
      description: CU session cookie, validated against the CU upstream on every request.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size (default 20, clamped to a maximum of 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip (default 0).
      schema: { type: integer, minimum: 0, default: 0 }
...
    SkillId:
      name: skillId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or rejected session cookie
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: >
        Normalized error envelope returned by every error response. "error" is a
        human-readable message (not for programmatic use); "code" is a stable,
        machine-readable identifier clients may branch on.
      properties:
        error: { type: string, example: "announcement not found" }
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - already_exists
            - invalid_reference
            - payload_too_large
            - upstream_error
            - unavailable
            - internal_error
      required: [error, code]
    PageMeta:
      type: object
      properties:
        total: { type: integer, description: Total number of matching items }
        limit: { type: integer }
        offset: { type: integer }
      required: [total, limit, offset]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        photoUrl:
          type: string
          nullable: true
          description: Presigned, time-limited download URL for the user's photo.
        firstName: { type: string, maxLength: 31 }
        lastName: { type: string, maxLength: 31 }
        bio: { type: string, nullable: true, maxLength: 255 }
        birthdate: { type: string, format: date, example: "2000-01-31" }
        speciality: { type: string, nullable: true, maxLength: 63 }
        status: { type: string, nullable: true, maxLength: 127, description: Resolved status content }
        role: { type: string, enum: [user, admin] }
      required: [id, firstName, lastName, birthdate, role]
...
    Skill:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 63 }
      required: [id, name]
...
    ProfileData:
      type: object
      description: Profile fields (partial update; omitted fields keep their current value)
      properties:
        bio: { type: string, nullable: true, maxLength: 255 }
        speciality: { type: string, nullable: true, maxLength: 63 }
        statusId: { type: integer, nullable: true }
...
    SkillPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Skill" } }
          required: [data]
...